requirements.txt
torch
diffusers
transformers
accelerate
safetensors
numpy
opencv-python
pillow
pyyaml
packaging
openai
tqdm

